+++ src/categoriasQuestoes.ts
@@ -1,0 +1,29 @@
+import { ICategoriasQuestoes } from './interface';
+
+export const categoriasQuestoes: ICategoriasQuestoes[] = [
+    {
+        id: 1,
+        name: 'BI',
+        order: 1
+    },
+    {
+        id: 2,
+        name: 'Cloud',
+        order: 2
+    },
+    {
+        id: 3,
+        name: 'Linux',
+        order: 3
+    },
+    {
+        id: 4,
+        name: 'Phyton',
+        order: 4
+    },
+    {
+        id: 5,
+        name: 'Redes',
+        order: 5
+    }
+];