+++ src/screens/login/loginCreate.tsx
@@ -54,7 +54,7 @@ function LoginCreate(): ReactElement {
             if (routeToRedirect) {
                 navigation.dispatch(CommonActions.navigate({ name: routeToRedirect, params: routeParams }));
             } else {
-                navigation.dispatch(CommonActions.navigate({ name: 'Usuario' }));
+                navigation.dispatch(CommonActions.navigate({ name: 'Categorias Questões' }));
             }
         }